Fix nested set database issue and optimize storage package
- Resolve 'unable to encode 0x0' error by changing nested set fields from uint64 to int64
- Update all entities (Media, Comment, Menu, Organization, Taxonomy) to use int64 for nested set fields
- Update nested set helper functions to use int64 consistently
- Fix all repository adapters to handle int64 nested set values
- Update HTTP request/response DTOs to use int64 for nested set fields
- Move storage package from internal/infrastructure/storage to pkg/storage
- Optimize storage package with better error handling and health checks
- Add comprehensive storage documentation and troubleshooting guides
- Fix database migration constraints and seed data
- Update service layer to handle int64 nested set values
- Ensure type consistency across all layers of the application

+++ internal/infrastructure/adapters/comment_repository.go
@@ -67,8 +67,8 @@ func NewPostgresCommentRepository(db *pgxpool.Pool, redisClient redis.Cmdable) r
 // Returns:
 //   - error: Any error that occurred during the database operation
 func (r *PostgresCommentRepository) Create(ctx context.Context, comment *entities.Comment) error {
-	// Calculate nested set values using the nested set manager
-	nestedSetValues, err := r.nestedSetManager.CreateNode(ctx, "comments", comment.ParentID, 0)
+	// Calculate nested set values
+	nestedSetValues, err := r.nestedSetManager.CreateNode(ctx, "comments", comment.ParentID, int64(0))
 	if err != nil {
 		return fmt.Errorf("failed to calculate nested set values: %w", err)
 	}
@@ -139,7 +139,7 @@ func (r *PostgresCommentRepository) Create(ctx context.Context, comment *entitie
 		return fmt.Errorf("failed to insert content: %w", err)
 	}
 
-	// Update the comment entity with the calculated nested set values
+	// Assign nested set values to comment entity
 	comment.RecordLeft = &nestedSetValues.Left
 	comment.RecordRight = &nestedSetValues.Right
 	comment.RecordDepth = &nestedSetValues.Depth
